requirements.txt
pyaml==21.10.1
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
urllib3==1.26.14
